Update RGraph.common.sheets.js
Update the Google Sheets import utility to work with the v4 Google API. Instead of dynamically adding a <script> tag it now usesan AJAX call. As such the AJAX methods have been added along with a copy of the getArgs() function too. This so that the object works without the RGRaph core library.